resize_drag: consolidate handleSegmentResizeEnd()
